# _example/api/order/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Service",
    "version": "v0.1.0"
  },
  "host": "localhost",
  "basePath": "/",
  "paths": {
    "/api/v1/order-service/order": {
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Service"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "payload",
            "in": "body",
            "schema": {
              "properties": {
                "order_info": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "item_id": {
                        "type": "string"
                      },
                      "quantity": {
                        "type": "integer",
                        "format": "int64"
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int64"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/order-service/order/cancel": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Service"
        ],
        "operationId": "cancel-order",
        "parameters": [
          {
            "name": "payload",
            "in": "body",
            "schema": {
              "properties": {
                "order_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int64"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/order-service/order/detail": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Service"
        ],
        "operationId": "order-detail",
        "parameters": [
          {
            "type": "string",
            "name": "order_id",
            "in": "query"
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int64"
                },
                "message": {
                  "type": "string"
                },
                "order_info": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "item_id": {
                        "type": "string"
                      },
                      "quantity": {
                        "type": "integer",
                        "format": "int64"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}